Make animation smoother
This granularity results in 30+fps, which looks smooth even for a 10
second duration and doesn't seem to use a lot of CPU (tested with
Chrome).
"transition: d 0.1s" or similar is not a good option, because the arc
would distort if a large-ish angle needs to be transitioned, as it would
be for a short duration. Either way, it doesn't seem to use much less
CPU than the simple interval.