Improve wobble motion
This makes the motion less predictable, giving x and y a slightly
different speed in an otherwise normal circular motion. This most
importantly doesn't vary the radius so much on the y component, which
together with the amplitude previously resulted in jerky, rapid motion
whenever amplitude and t-shift aligned just so.

This way still has some faster/slower moments, but mostly due to how the
different shifts interact, and it looks random but never crazy.